# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Email Microservice API",
    "description": "API untuk Email Microservice yang mendukung pengiriman email, manajemen template, dan pelacakan email.",
    "version": "1.0.0",
    "contact": {
      "name": "Tim API Email Microservice",
      "email": "api-support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/tracking/{email_id}": {
      "get": {
        "tags": ["Tracking"],
        "summary": "Mendapatkan tracking email",
        "description": "Mendapatkan data tracking untuk email tertentu",
        "operationId": "getTracking",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "description": "ID email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data tracking berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Email tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TrackingEvent": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "queued",
              "sent",
              "delivered",
              "opened",
              "clicked",
              "bounced",
              "complained",
              "failed"
            ],
            "description": "Jenis event"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Waktu event"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Metadata event"
          }
        }
      },
      "TrackingResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "emailId": {
                "type": "string",
                "description": "ID email"
              },
              "status": {
                "type": "string",
                "enum": [
                  "pending",
                  "sending",
                  "delivered",
                  "failed",
                  "bounced",
                  "rejected",
                  "complained"
                ],
                "description": "Status email"
              },
              "events": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TrackingEvent"
                },
                "description": "Daftar event tracking"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "description": "Nama field yang error"
                    },
                    "message": {
                      "type": "string",
                      "description": "Deskripsi error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    }
  }
}